部分逻辑修改

+++ salary-pojo/src/main/java/com/dfd/dto/ItemPlanDTO.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * @author yy
@@ -37,41 +38,41 @@ public class ItemPlanDTO implements Serializable {
      * 设计阶段系数
      */
     @ApiModelProperty(value = "设计阶段系数", name = "designCoefficient")
-    private Integer designCoefficient;
+    private BigDecimal designCoefficient;
 
     /**
      * 采购阶段系数
      */
     @ApiModelProperty(value = "采购阶段系数", name = "purchaseCoefficient")
-    private Integer purchaseCoefficient;
+    private BigDecimal purchaseCoefficient;
 
     /**
      * 制造阶段系数
      */
     @ApiModelProperty(value = "制造阶段系数", name = "manufactureCoefficient")
-    private Integer manufactureCoefficient;
+    private BigDecimal manufactureCoefficient;
 
     /**
      * 安装调试阶段系数
      */
     @ApiModelProperty(value = "安装调试阶段系数", name = "installationCoefficient")
-    private Integer installationCoefficient;
+    private BigDecimal installationCoefficient;
 
     /**
      * 安装验收阶段系数
      */
     @ApiModelProperty(value = "安装验收阶段系数", name = "inspectionCoefficient")
-    private Integer inspectionCoefficient;
+    private BigDecimal inspectionCoefficient;
 
     /**
      * 终验收阶段系数
      */
     @ApiModelProperty(value = "终验收阶段系数", name = "finalCoefficient")
-    private Integer finalCoefficient;
+    private BigDecimal finalCoefficient;
 
     /**
      * 质保阶段系数
      */
     @ApiModelProperty(value = "质保阶段系数", name = "guaranteeCoefficient")
-    private Integer guaranteeCoefficient;
+    private BigDecimal guaranteeCoefficient;
 }
